# Conflicts:
#	src/contracts/contracts/partials/contractTypes/doormanTypes.ligo
#	src/contracts/contracts/partials/errors.ligo

+++ src/indexer/mavryk/handlers/on_vault_update_depositor.py
@@ -1,42 +1,47 @@
 
-from mavryk.types.vault.parameter.update_depositor import UpdateDepositorParameter
+from mavryk.types.vault.parameter.update_depositor import UpdateDepositorParameter, AllowanceItem as Any, AllowanceItem1 as Whitelist
 from dipdup.context import HandlerContext
 from dipdup.models import Transaction
-from mavryk.types.vault.storage import VaultStorage, Depositor as Any, Depositor1 as Whitelist
+from mavryk.types.vault.storage import VaultStorage
 import mavryk.models as models
 
 async def on_vault_update_depositor(
     ctx: HandlerContext,
     update_depositor: Transaction[UpdateDepositorParameter, VaultStorage],
 ) -> None:
-    
+
     # Get operation info
     vault_address       = update_depositor.data.target_address
-    depositors          = update_depositor.storage.depositors
+    depositor           = update_depositor.parameter.allowance
     allowance_type      = models.VaultAllowance.ANY
 
     # Update record
     vault               = await models.Vault.get(
         address = vault_address
     )
-    vault_depositors    = await models.VaultDepositor.filter(vault=vault).all()
-    for vault_depositor in vault_depositors:
-        await vault_depositor.delete()
 
-    if type(depositors) == Any:
-        allowance_type  = models.VaultAllowance.ANY
-    elif type(depositors) == Whitelist:
-        allowance_type  = models.VaultAllowance.WHITELIST
-        for depositor_address in depositors.whitelist:
-            depositor, _        = await models.MavrykUser.get_or_create(
-                address = depositor_address
-            )
-            await depositor.save()
-            vault_depositor, _  = await models.VaultDepositor.get_or_create(
-                vault       = vault,
-                depositor   = depositor
-            )
+    if type(depositor) == Any:
+        reset_whitelist     = not depositor.any
+        if reset_whitelist:
+            allowance_type      = models.VaultAllowance.WHITELIST
+            vault_depositors    = await models.VaultDepositor.filter(vault = vault).all()
+            for vault_depositor in vault_depositors:
+                await vault_depositor.delete()
+        else:
+            allowance_type      = models.VaultAllowance.ANY
+    elif type(depositor) == Whitelist:
+        allowance_type      = models.VaultAllowance.WHITELIST
+        depositor_address   = depositor.whitelist.address
+        add_depositor       = depositor.whitelist.bool
+        user                = await models.mavryk_user_cache.get(address=depositor_address)
+        vault_depositor, _  = await models.VaultDepositor.get_or_create(
+            vault       = vault,
+            depositor   = user
+        )
+        if add_depositor:
             await vault_depositor.save()
+        else:
+            await vault_depositor.delete()
 
     vault.allowance = allowance_type
     await vault.save()
